remove C/C++ building functionality from core
add task timing (by monkey patching task invocation in rake)
add API reference doc generation
move all C/C++ relevant functionality (including platform abstractions)
to gaudi-c module
