+++ Rakefile
@@ -5,13 +5,14 @@ require_relative('lib/gaudi/version')
 
 task :default => [:test]
 
-
-task :test do
+task :"test:gaudi" do
   require 'coveralls'
   Coveralls.wear!
   require 'minitest/autorun'
-  Rake::FileList["#{File.dirname(__FILE__)}/test/test_*.rb"].each do |test_file|
-    require_relative "test/#{test_file.pathmap('%n')}"
+  Rake::FileList["#{File.dirname(__FILE__)}/test/gaudi/test_*.rb"].each do |test_file|
+    require_relative "test/gaudi/#{test_file.pathmap('%n')}"
   end
 end
 # vim: syntax=ruby
+
+task :test=> [:"test:gaudi"]
