+++ lib/gaudi-c/helpers/artifacts.rb
