+++ lib/gaudi-c/helpers/components.rb
