+++ lib/gaudi-c/helpers/generators.rb
