+++ lib/gaudi-c/helpers/operations.rb
