+++ lib/gaudi-c/tasks/clean.rb
@@ -1,0 +1,8 @@
+namespace :clean do
+  desc "Removes the compilation command files"
+  task :breadcrumbs do
+    if $configuration
+      rm_rf(FileList[*$configuration.source_directories.pathmap('%p/**/*.{compile,assemble}')],:verbose=>false)
+    end
+  end
+end
