+++ lib/gaudi/version.rb
@@ -1,13 +1,14 @@
-#A bootstrap for creating a build system for C or C++ based projects using rake.
+#Gaudi is a collection of helpers and tools that together with a small set of conventions and rake
+#allows you to create complex build systems.
 module Gaudi
   #Gaudi follows SemVer even though it's not a gem
   module Version
     #Major version
-    MAJOR=0
+    MAJOR=1
     #Minor version
-    MINOR=12
+    MINOR=0
     #Tiny version
-    TINY=1
+    TINY=0
     #All-in-one
     STRING=[MAJOR,MINOR,TINY].join('.')
   end
