+++ test/gaudi-c/test_components.rb
