+++ test/gaudi-c/test_configuration.rb
