+++ test/gaudi-c/test_platform_operations.rb
