+++ test/gaudi-c/test_utilities.rb
